Refactor DBPedia-Exetractor
